Ready queue assignment
Emulates the ready queue of an operating system: Holds a list of
processes which are ready to run, and prioritizes them by urgency.
I used a heap tree to achieve this, so the most urgent job is always
element 0. And I used an array for the heap tree, for performance.